+++ src/Instance/InstanceEdit/InstanceEdit.test.js
@@ -46,6 +46,9 @@ const stripesStub = {
   logger: { log: noop },
   locale: 'en-US',
   plugins: {},
+  okapi: {
+    tenant: 'diku',
+  },
 };
 const mockInstance = {
   ...instance,