+++ src/constants.js
@@ -192,6 +192,7 @@ export const SYSTEM_USER_ID = '00000000-0000-0000-0000-000000000000';
 export const SYSTEM_USER_NAME = 'System';
 
 export const SINGLE_ITEM_QUERY_TEMPLATES = {
+  [queryIndexes.ITEMS_KEYWORD]: 'hrid=="%{query}" or id=="%{query}" or barcode=="%{query}"',
   [queryIndexes.ITEMS_BARCODE]: 'barcode=="%{query}"',
   [queryIndexes.ISBN]: 'isbn=="%{query}"',
   [queryIndexes.ISSN]: 'issn=="%{query}"',