+++ src/hooks/useHoldingItemsQuery.js
@@ -25,6 +25,7 @@ const useHoldingItemsQuery = (
   // sortMap contains not all item table's columns because sorting by some columns
   // is not implemented on BE yet
   const sortMap = {
+    'order': 'order/number',
     'barcode': 'barcode',
     'status': 'status.name',
     'copyNumber': 'copyNumber',
@@ -39,7 +40,9 @@ const useHoldingItemsQuery = (
       const sortQuery = options.searchParams.sortBy;
       const sortDirection = sortQuery.startsWith('-') ? 'descending' : 'ascending';
       const sortOrder = sortQuery.replace(/^-/, '');
-      const newSortBy = sortMap[sortOrder] ? `${sortMap[sortOrder]}/sort.${sortDirection}` : sortBy;
+      const newSortBy = sortMap[sortOrder]
+        ? `${sortMap[sortOrder]}/sort.${sortDirection}`
+        : sortBy;
 
       setSortBy(newSortBy);
     }